+++ src/themes/_generator/transform.ts
@@ -9,7 +9,7 @@ import * as transforms from "./tokens/transforms";
 import { variablesTemplate, mediaTemplate } from "./tokens/css";
 import { generateUnits } from "./tokens/unit/utilities/generate";
 import generateMetaColors from "./tokens/color/utilities/generateMetaColors";
-import { formatHex } from "culori/fn";
+import { formatHex8, formatHex } from "culori/fn";
 import { Token } from "./tokens/color/color.types";
 
 const transform = (name: string, definition: PassedThemeDefinition, options: T.PrivateOptions) => {
@@ -26,10 +26,16 @@ const transform = (name: string, definition: PassedThemeDefinition, options: T.P
 		const next = { ...token } as Token;
 		if (!token) return res;
 		if (!token.hex && token.oklch) {
-			next.hex = formatHex({ ...token.oklch, mode: "oklch" });
+			next.hex =
+				token.oklch.alpha !== undefined
+					? formatHex8({ ...token.oklch, mode: "oklch" })
+					: formatHex({ ...token.oklch, mode: "oklch" });
 		}
 		if (!token.hexDark && token.oklchDark) {
-			next.hexDark = formatHex({ ...token.oklchDark, mode: "oklch" });
+			next.hexDark =
+				token.oklchDark.alpha !== undefined
+					? formatHex8({ ...token.oklchDark, mode: "oklch" })
+					: formatHex({ ...token.oklchDark, mode: "oklch" });
 		}
 
 		return { ...res, [key]: next };